Фетч в мейн.жс
