+++ js/user-form.js
@@ -1,75 +1,44 @@
-const form = document.querySelector('.img-upload__form');
+import {isEscapeKey} from './utilites.js';
 
-const pristine = new Pristine(form, {
-  classTo: 'img-upload__field-wrapper',
-  errorTextParent: 'img-upload__field-wrapper',
-  errorTextClass: 'error-text',
-});
+const uploadImage = document.querySelector('#upload-file');
+const uploadImageForm = document.querySelector('.img-upload__overlay');
+const uploadImageDescription = document.querySelector('.text__description');
+const uploadImageHashtags = document.querySelector('.text__hashtags');
 
-// валидация хэштегов
+const uploadImageClose = document.querySelector('#upload-cancel');
 
-const hashtagsField = document.querySelector('.text__hashtags');
-const re = /^#[A-Za-zА-Яа-яЁё0-9]{1,19}$/;
+function openForm () {
+  uploadImageForm.classList.remove('hidden');
+  document.body.classList.add('modal-open');
+}
 
-const validateHashtagValue = function (value) {
-  const array = value.split(' ');
-  for (const arrayElement of array) {
-    if (re.test(arrayElement) || arrayElement === '') {
-      return true;
-    }
-  } return false;
-};
+function closeForm () {
+  uploadImageForm.classList.add('hidden');
+  document.body.classList.remove('modal-open');
+  uploadImage.value = '';
+  uploadImageDescription.value ='';
+  uploadImageHashtags.value = '';
+  document.removeEventListener('keydown', onFormEscKeydown);
+}
 
-const validateHashtagLength = function (value) {
-  const maxHastagLength = 20;
-  const array = value.split(' ');
-  for (const arrayElement of array) {
-    if (arrayElement.length <= maxHastagLength) {
-      return true;
-    }
-  } return false;
-};
-
-const validateHashtagsQuantity = function (value) {
-  const array = value.split(' ');
-  const maxHastagsQuantity = 5;
-  return array.length <= maxHastagsQuantity;
-};
-
-const validateNotSameHashtags = function (value) {
-  const array = value.split(' ');
-  for (let i = 0; i < array.length; i++) {
-    for (let j = i + 1; j < array.length; j++) {
-      if (array[i] === array[j]) {
-        return false;
-      }
-    }
-  } return true;
-};
-
-const validateNotSameInLowerCase = function (value) {
-  const array = value.split(' ');
-  for (let i = 0; i < array.length; i++) {
-    for (let j = i + 1; j < array.length; j++) {
-      if (array[i].toLowerCase() === array[j].toLowerCase()) {
-        return false;
-      }
-    }
-  } return true;
-};
-
-pristine.addValidator(hashtagsField, validateHashtagLength, 'максимальная длина одного хэш-тега 20 символов, включая решётку');
-pristine.addValidator(hashtagsField, validateHashtagsQuantity, 'нельзя указать больше пяти хэш-тегов');
-pristine.addValidator(hashtagsField, validateHashtagValue, 'хэш-тег начинается с символа # (решётка), строка после решётки должна состоять из букв и чисел и не может содержать пробелы, спецсимволы (#, @, $ и т. п.), символы пунктуации (тире, дефис, запятая и т. п.), эмодзи и т. д., хеш-тег не может состоять только из одной решётки', false);
-pristine.addValidator(hashtagsField, validateNotSameHashtags, 'один и тот же хэш-тег не может быть использован дважды');
-pristine.addValidator(hashtagsField, validateNotSameInLowerCase, 'хэш-теги нечувствительны к регистру: #ХэшТег и #хэштег считаются одним и тем же тегом');
-
-form.addEventListener('submit', (evt) => {
-  const isValid = pristine.validate();
-  if (!isValid) {
-    evt.preventDefault();
-  }
-  else {
+function onFormEscKeydown (evt) {
+  if (uploadImageDescription === document.activeElement || uploadImageHashtags === document.activeElement) {
     return evt;
+  } else {
+    if (isEscapeKey(evt)) {
+      evt.preventDefault();
+      closeForm ();
+    }
   }
+}
+
+uploadImage.addEventListener('change', () => {
+  openForm();
+  document.addEventListener('keydown', onFormEscKeydown);
 });
+
+uploadImageClose.addEventListener('click', () => {
+  closeForm();
+});
+
+export {closeForm, openForm};
